Added method to read in continuous data.
